requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
pyyaml>=6.0
tensorboard>=2.13.0
soundfile>=0.12.0
jiwer>=3.0.0
pyworld>=0.3.2
encodec>=0.1.1
tqdm>=4.65.0
omegaconf>=2.3.0
torchcodec==0.9.1
